Boss AI structure implemented
I have added a series of scripts and classes which will work together to act as the AI for the bosses.  Next I will need to implement the heuristics, the learning algorithms, and the action space.  If you are curious as to how the  AI works feel free to reach out to me.