Minor Code Refactor
Shifted around the boss code in order to gather it all up in one place.  This will be my last update of the night, tomorrow I have to make a few more heuristics, then design the algorithms to generate bosses and decision trees as well as modify existing trees using a memory.  Finally I will need to design all the weapons so the artists can begin creating them.